fix(api): move bulk-locations route to own root to prevent route collisions
